Implement DirectAvoidingLoopbackFetcher
This is an absolutely *wild* workaround for the fact that libp2p won't let us dial to ourselves.

gomatrixserverlib's DirectKeyFetcher expects to always be able to make a HTTP request to find server keys, even to ourselves. This type of loopback dial doesn't work on the P2P branch and therefore server federation breaks as it can't verify event signatures properly.

This implements something that stops us from having to make HTTP requests for our own keys. Phew.